Refactor: config now a package, not a single file
